fixed absolute paths
