+++ .github/workflows/test.yml
@@ -29,18 +29,19 @@ jobs:
         python-version: ${{ matrix.python-version }}
         cache: "pip"
         cache-dependency-path: pyproject.toml
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install -e ".[dev-test]"
+
     - name: Checkout qudi core
       run: |
         cd ..
         git clone https://github.com/prithviulm/qudi-core.git
         cd qudi-core
+        python -m pip install --upgrade pip
         pip install -e ".[dev-test]"
         cd ../qudi-iqo-modules
 
+    - name: Install dependencies iqo modules
+      run: |
+        pip install -e ".[dev-test]"
     - name: Test with pytest
       run: |
         pytest -s
